Change names of PWHL teams in Qualtrics
Use "PWHL Boston" instead of "Boston", "PWHL Minnesota" instead of "Minnesota," etc. This reads better to avoid people thinking that the questions are asking about the locations rather than the teams. E.g., "When someone praises New York, it feels like a personal compliment" could refer to the city or the team.