Allowed for non-rival object to be null
If someone named all the teams as rivals, the objNonRival object would be null. Attempted to read properties of it crash the code and block the user from clicking the 'Next' button.

The fix uses the optional chaining (?.) and nullish coalescing (??) operators to allow for this.
```js
objNonRival.entID --becomes--> objNonRival?.entID ?? -99
objNonRival.entityName --becomes--> objNonRival?.entityName ?? 'non-rival team'
```